requirements.txt
elasticsearch>=2.0.0
elasticsearch-curator>=3.4.0